add revoke token service spec
