--- app/services/api_clients/revoke_token_service.rb
@@ -1,17 +1,0 @@
-module ApiClients
-  class RevokeTokenService
-    include ServicePattern
-
-    def initialize(api_client:)
-      @api_client = api_client
-    end
-
-    def call
-      api_client.current_authentication_token.revoke!
-    end
-
-  private
-
-    attr_reader :api_client
-  end
-end
