+++ db/schema.rb
@@ -59,9 +59,9 @@ ActiveRecord::Schema[8.1].define(version: 2026_03_19_161428) do
     t.datetime "created_at", null: false
     t.uuid "created_by_id", null: false
     t.datetime "discarded_at"
-    t.string "name", null: false
+    t.citext "name", null: false
     t.datetime "updated_at", null: false
-    t.index "created_by_id, lower((name)::text)", name: "index_api_clients_on_created_by_and_lower_name"
+    t.index "created_by_id, lower((name)::text)", name: "index_api_clients_on_created_by_and_lower_name", unique: true, where: "(discarded_at IS NULL)"
     t.index ["discarded_at"], name: "index_api_clients_on_discarded_at"
   end
